+++ backend/integration_tests/src/client/registry.rs
@@ -30,19 +30,23 @@ pub mod happy_path {
 
         assert!(matches!(response, expand_onto_subnet::Response::Success));
 
-        for _ in 0..40 {
+        for i in 1..200 {
             env.advance_time(Duration::from_secs(1));
             env.tick();
+
+            if i % 10 == 0 {
+                let subnets::Response::Success(subnets) = super::subnets(env, sender, registry_canister_id, &Empty {});
+
+                if let Some(subnet) = subnets.into_iter().find(|s| s.subnet_id == subnet_id) {      
+                    env.add_cycles(subnet.local_user_index, INIT_CYCLES_BALANCE);
+                    env.add_cycles(subnet.local_group_index, INIT_CYCLES_BALANCE);
+                    env.add_cycles(subnet.notifications_canister, INIT_CYCLES_BALANCE);
+
+                    return subnet;
+                }
+            }
         }
 
-        let subnets::Response::Success(subnets) = super::subnets(env, sender, registry_canister_id, &Empty {});
-
-        let subnet = subnets.last().unwrap().clone();
-
-        env.add_cycles(subnet.local_user_index, INIT_CYCLES_BALANCE);
-        env.add_cycles(subnet.local_group_index, INIT_CYCLES_BALANCE);
-        env.add_cycles(subnet.notifications_canister, INIT_CYCLES_BALANCE);
-
-        subnet
+        panic!("Failed to expand onto new subnet")
     }
 }